app/requirements.txt
Flask
pymongo
pylint
python-dotenv